Move `hikari.impl.bot` to `hikari.impl.gateway_bot`
