+++ changes/1576.deprecation.md
@@ -1,0 +1,1 @@
+`hikari.impl.bot` moved to `hikari.impl.gateway_bot`.
