+++ hikari/impl/__init__.pyi
@@ -1,13 +1,13 @@
 # DO NOT MANUALLY EDIT THIS FILE!
 # This file was automatically generated by `nox -s generate-stubs`
 
-from hikari.impl.bot import *
 from hikari.impl.buckets import *
 from hikari.impl.cache import *
 from hikari.impl.config import *
 from hikari.impl.entity_factory import *
 from hikari.impl.event_manager import *
 from hikari.impl.event_manager_base import *
+from hikari.impl.gateway_bot import *
 from hikari.impl.interaction_server import *
 from hikari.impl.rate_limits import *
 from hikari.impl.rest import *
