+++ tests/hikari/test_presences.py
@@ -27,13 +27,13 @@ from hikari import files
 from hikari import presences
 from hikari import snowflakes
 from hikari import urls
-from hikari.impl import bot
+from hikari.impl import gateway_bot
 from hikari.internal import routes
 
 
 @pytest.fixture()
 def mock_app():
-    return mock.Mock(spec_set=bot.GatewayBot)
+    return mock.Mock(spec_set=gateway_bot.GatewayBot)
 
 
 class TestActivityAssets:
